+++ temba/airtime/tests.py
@@ -43,8 +43,7 @@ class AirtimeEventTest(TembaTest):
             model_obj_response = self.airtime.response
 
             response = AirtimeTransfer.post_transferto_api_response('login_acc', 'token', action='ping')
-            self.assertEqual(response.status_code, 200)
-            self.assertEqual(response.content, "foo=allo\r\nbar=1,2,3\r\n")
+            self.assertContains(response, "foo=allo\r\nbar=1,2,3\r\n")
 
             self.assertEqual(mock_post.call_count, 1)
             self.assertEqual('https://airtime.transferto.com/cgi-bin/shop/topup', mock_post.call_args_list[0][0][0])
@@ -65,8 +64,7 @@ class AirtimeEventTest(TembaTest):
 
             response = AirtimeTransfer.post_transferto_api_response('login_acc', 'token', airtime_obj=self.airtime,
                                                                     action='ping')
-            self.assertEqual(response.status_code, 200)
-            self.assertEqual(response.content, "foo=allo\r\nbar=1,2,3\r\n")
+            self.assertContains(response, "foo=allo\r\nbar=1,2,3\r\n")
             self.assertEqual(mock_post.call_count, 1)
             self.assertEqual('https://airtime.transferto.com/cgi-bin/shop/topup', mock_post.call_args_list[0][0][0])
             mock_args = mock_post.call_args_list[0][0][1]
@@ -90,8 +88,7 @@ class AirtimeEventTest(TembaTest):
 
         with self.settings(SEND_AIRTIME=True):
             response = self.airtime.get_transferto_response(action='command')
-            self.assertEqual(200, response.status_code)
-            self.assertEqual(response.content, "foo=allo\r\nbar=1,2,3\r\n")
+            self.assertContains(response, "foo=allo\r\nbar=1,2,3\r\n")
 
             mock_post_transferto.assert_called_once_with('mylogin', 'api_token', airtime_obj=self.airtime,
                                                          action='command')