+++ temba/channels/tests.py
@@ -13,7 +13,7 @@ import pytz
 import six
 import time
 
-from django.utils.encoding import force_text
+from django.utils.encoding import force_text, force_bytes
 from six.moves.urllib.parse import quote, urlencode
 import uuid
 
@@ -28,7 +28,7 @@ from django.test.utils import override_settings
 from django.utils import timezone
 from django.template import loader
 from django_redis import get_redis_connection
-from mock import patch
+from mock import patch, ANY
 from smartmin.tests import SmartminTest
 from temba.api.models import WebHookEvent
 from temba.contacts.models import Contact, ContactGroup, ContactURN, URN, TEL_SCHEME, TWITTER_SCHEME, EXTERNAL_SCHEME, \
@@ -574,7 +574,7 @@ class ChannelTest(TembaTest):
 
             # sign the request
             key = str(channel.secret) + str(ts)
-            signature = hmac.new(key=key, msg=bytes(post_data), digestmod=hashlib.sha256).digest()
+            signature = hmac.new(key=force_bytes(key), msg=force_bytes(post_data), digestmod=hashlib.sha256).digest()
 
             # base64 and url sanitize
             signature = quote(base64.urlsafe_b64encode(signature))
@@ -1281,7 +1281,6 @@ class ChannelTest(TembaTest):
             ]
             response = self.client.post(plivo_search_url, dict(country='US', area_code=''), follow=True)
 
-            self.assertEqual(response.status_code, 200)
             self.assertContains(response, 'Bad request')
 
     def test_release(self):
@@ -2544,9 +2543,8 @@ class ExternalTest(TembaTest):
         data = {'from': '5511996458779', 'text': 'Hello World!', 'date': '2012-04-23T18:25:43Z'}
         response = self.client.post(callback_url, data)
 
-        self.assertEqual(400, response.status_code)
-        self.assertEqual("Bad parameter error: time data '2012-04-23T18:25:43Z' "
-                         "does not match format '%Y-%m-%dT%H:%M:%S.%fZ'", response.content)
+        self.assertContains(response, "Bad parameter error: time data '2012-04-23T18:25:43Z' "
+                            "does not match format '%Y-%m-%dT%H:%M:%S.%fZ'", status_code=400)
         self.assertFalse(Msg.objects.all())
 
     def test_receive_external(self):
@@ -2590,9 +2588,9 @@ class ExternalTest(TembaTest):
                 mock.return_value = MockResponse(200, "Sent")
                 Channel.send_message(dict_to_struct('MsgStruct', msg.as_task_json()))
                 self.assertEqual(mock.call_args[0][0], 'http://foo.com/send')
-                self.assertEqual(mock.call_args[1]['data'], 'id=%d&text=Test+message&to=%%2B250788383383&to_no_plus=250788383383&'
-                                                            'from=%%2B250788123123&from_no_plus=250788123123&'
-                                                            'channel=%d' % (msg.id, self.channel.id))
+                self.assertEqual(mock.call_args[1]['data'], force_bytes('id=%d&text=Test+message&to=%%2B250788383383&to_no_plus=250788383383&'
+                                                                        'from=%%2B250788123123&from_no_plus=250788123123&'
+                                                                        'channel=%d' % (msg.id, self.channel.id)))
 
         self.channel.config = {Channel.CONFIG_SEND_URL: 'http://foo.com/send',
                                Channel.CONFIG_SEND_BODY: '{ "text": {{text}}, "to": {{to_no_plus}} }',
@@ -2607,7 +2605,7 @@ class ExternalTest(TembaTest):
                 mock.return_value = MockResponse(200, "Sent")
                 Channel.send_message(dict_to_struct('MsgStruct', msg.as_task_json()))
                 self.assertEqual(mock.call_args[0][0], 'http://foo.com/send')
-                self.assertEqual(mock.call_args[1]['data'], '{ "text": "Test message", "to": "250788383383" }')
+                self.assertEqual(mock.call_args[1]['data'], force_bytes('{ "text": "Test message", "to": "250788383383" }'))
                 self.assertEqual(mock.call_args[1]['headers']['Content-Type'], "application/json")
 
         self.channel.config = {Channel.CONFIG_SEND_URL: 'http://foo.com/send',
@@ -2625,7 +2623,7 @@ class ExternalTest(TembaTest):
                 mock.return_value = MockResponse(200, "Sent")
                 Channel.send_message(dict_to_struct('MsgStruct', msg.as_task_json()))
                 self.assertEqual(mock.call_args[0][0], 'http://foo.com/send')
-                self.assertEqual(mock.call_args[1]['data'], 'text=' + msg.text + '&to=250788383383')
+                self.assertEqual(mock.call_args[1]['data'], force_bytes('text=' + msg.text + '&to=250788383383'))
 
         self.channel.config = {Channel.CONFIG_SEND_URL: 'http://foo.com/send',
                                Channel.CONFIG_SEND_BODY: '<msg><text>{{text}}</text><to>{{to_no_plus}}</to></msg>',
@@ -2643,7 +2641,7 @@ class ExternalTest(TembaTest):
                 mock.return_value = MockResponse(200, "Sent")
                 Channel.send_message(dict_to_struct('MsgStruct', msg.as_task_json()))
                 self.assertEqual(mock.call_args[0][0], 'http://foo.com/send')
-                self.assertEqual(mock.call_args[1]['data'], '<msg><text>التوطين</text><to>250788383383</to></msg>'.encode('utf8'))
+                self.assertEqual(mock.call_args[1]['data'], force_bytes('<msg><text>التوطين</text><to>250788383383</to></msg>'))
                 self.assertEqual(mock.call_args[1]['headers']['Content-Type'], Channel.CONTENT_TYPES[Channel.CONTENT_TYPE_XML])
 
     @override_settings(SEND_MESSAGES=True)
@@ -2662,7 +2660,7 @@ class ExternalTest(TembaTest):
             self.assertEqual(WIRED, msg.status)
             self.assertTrue(msg.sent_on)
 
-            self.assertIn("text=Test+message", mock.call_args[1]['data'])
+            self.assertIn(force_bytes("text=Test+message"), mock.call_args[1]['data'])
 
             self.clear_cache()
 
@@ -2732,7 +2730,7 @@ class ExternalTest(TembaTest):
             self.assertEqual(WIRED, msg.status)
             self.assertTrue(msg.sent_on)
 
-            self.assertIn("text=Test+message%0Ahttps%3A%2F%2Fexample.com%2Fattachments%2Fpic.jpg", mock.call_args[1]['data'])
+            self.assertIn(force_bytes("text=Test+message%0Ahttps%3A%2F%2Fexample.com%2Fattachments%2Fpic.jpg"), mock.call_args[1]['data'])
 
             self.clear_cache()
 
@@ -3985,8 +3983,7 @@ class MacrokioskTest(TembaTest):
                                args=['receive', self.channel.uuid]) + "?" + encoded_message
         response = self.client.get(callback_url)
 
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.content, "-1")
+        self.assertContains(response, "-1")
 
         # load our message
         msg = Msg.objects.get()
@@ -4014,8 +4011,7 @@ class MacrokioskTest(TembaTest):
                                args=['receive', self.channel.uuid]) + "?" + encoded_message
         response = self.client.get(callback_url)
 
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.content, "-1")
+        self.assertContains(response, "-1")
 
         # load our message
         msg = Msg.objects.get()
@@ -4683,8 +4679,7 @@ class DartMediaTest(TembaTest):
         callback_url = reverse('handlers.dartmedia_handler', args=['received', self.channel.uuid]) + "?" + encoded_message
         response = self.client.get(callback_url)
 
-        self.assertEqual(401, response.status_code)
-        self.assertEqual(response.content, "Parameters message, original and sendto should not be null.")
+        self.assertContains(response, "Parameters message, original and sendto should not be null.", status_code=401)
 
         # all needed params
         data = {
@@ -6457,17 +6452,22 @@ class TwitterTest(TembaTest):
 
             Channel.send_message(dict_to_struct('MsgStruct', msg.as_task_json()))
 
+            data = json.dumps(dict(event=dict(message_create=dict(message_data=dict(
+                text='Hello, world!',
+                quick_reply=dict(
+                    type='options',
+                    options=[dict(label='Yes'), dict(label='No')]
+                )),
+                target=dict(recipient_id='10002')),
+                type='message_create')
+            ))
+
             mock.assert_called_with('https://api.twitter.com/1.1/direct_messages/events/new.json',
                                     files=None,
-                                    data=json.dumps(dict(event=dict(message_create=dict(message_data=dict(
-                                        text='Hello, world!',
-                                        quick_reply=dict(
-                                            type='options',
-                                            options=[dict(label='Yes'), dict(label='No')]
-                                        )),
-                                        target=dict(recipient_id='10002')),
-                                        type='message_create')
-                                    )))
+                                    data=ANY)
+
+            args, kwargs = mock.call_args
+            six.assertCountEqual(self, data, kwargs.get('data'))
 
             msg.refresh_from_db()
             self.assertEqual(msg.status, WIRED)
@@ -7122,7 +7122,7 @@ class JasminTest(TembaTest):
         def assertStatus(sms, dlvrd, err, assert_status):
             data['dlvrd'] = dlvrd
             data['err'] = err
-            response = self.client.post(reverse('handlers.jasmin_handler', args=['status', self.channel.uuid]), data)
+            response = self.client.post(reverse('courier.js', args=[self.channel.uuid, 'status']), data)
             self.assertEqual(200, response.status_code)
             sms = Msg.objects.get(pk=sms.id)
             self.assertEqual(assert_status, sms.status)
@@ -7141,11 +7141,10 @@ class JasminTest(TembaTest):
             'content': gsm7.encode("événement")[0],
             'id': 'external1'
         }
-        callback_url = reverse('handlers.jasmin_handler', args=['receive', self.channel.uuid])
+        callback_url = reverse('courier.js', args=[self.channel.uuid, 'receive'])
         response = self.client.post(callback_url, data)
 
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.content, "ACK/Jasmin")
+        self.assertContains(response, "ACK/Jasmin")
 
         # load our message
         msg = Msg.objects.get()
@@ -7293,8 +7292,7 @@ class JunebugTest(JunebugTestMixin, TembaTest):
                                args=['event', self.channel.uuid])
         response = self.client.post(delivery_url, data=json.dumps({}),
                                     content_type='application/json')
-        self.assertEqual(400, response.status_code)
-        self.assertIn('Missing one of', response.content)
+        self.assertContains(response, 'Missing one of', status_code=400)
 
     def test_status(self):
         # ok, what happens with an invalid uuid?
@@ -7345,11 +7343,10 @@ class JunebugTest(JunebugTestMixin, TembaTest):
                     args=['event', self.channel.uuid]),
             data=json.dumps(data),
             content_type='application/json')
-        self.assertEqual(400, response.status_code)
-        self.assertEqual(
-            response.content,
+        self.assertContains(
+            response,
             "Message with external id of '%s' not found" % (
-                data['message_id'],))
+                data['message_id'],), status_code=400)
 
     def test_status_with_auth(self):
         config = self.channel.config
@@ -7401,8 +7398,7 @@ class JunebugTest(JunebugTestMixin, TembaTest):
                                args=['inbound', self.channel.uuid])
         response = self.client.post(callback_url, json.dumps({}),
                                     content_type='application/json')
-        self.assertEqual(400, response.status_code)
-        self.assertIn('Missing one of', response.content)
+        self.assertContains(response, 'Missing one of', status_code=400)
 
     def test_receive(self):
         data = self.mk_msg(content="événement")
@@ -7635,8 +7631,7 @@ class MbloxTest(TembaTest):
             Msg.objects.filter(id=msg.id).update(status=WIRED)
             data['status'] = status
             response = self.client.post(delivery_url, json.dumps(data), content_type="application/json")
-            self.assertEqual(200, response.status_code)
-            self.assertEqual(response.content, "SMS Updated: %d" % msg.id)
+            self.assertContains(response, "SMS Updated: %d" % msg.id)
             msg = Msg.objects.get(pk=msg.id)
             self.assertEqual(assert_status, msg.status)
 
@@ -7661,8 +7656,7 @@ class MbloxTest(TembaTest):
 
         msg = Msg.objects.get()
 
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.content, "SMS Accepted: %d" % msg.id)
+        self.assertContains(response, "SMS Accepted: %d" % msg.id)
 
         # load our message
         self.assertEqual(msg.contact.get_urn(TEL_SCHEME).path, "+12067799294")
@@ -8009,8 +8003,7 @@ class FacebookTest(TembaTest):
         data = json.loads(FacebookTest.TEST_INCOMING)
         data['entry'][0]['messaging'][0] = json.loads(optin)
         response = self.client.post(callback_url, json.dumps(data), content_type='application/json')
-        self.assertEqual(200, response.status_code)
-        self.assertEqual('Msg Ignored for recipient id: PAGE_ID', response.content)
+        self.assertContains(response, 'Msg Ignored for recipient id: PAGE_ID')
 
         response = self.client.post(callback_url, json.dumps(data).replace('PAGE_ID', '1234'), content_type='application/json')
         self.assertEqual(200, response.status_code)
@@ -8022,8 +8015,7 @@ class FacebookTest(TembaTest):
         # try an invalid optin (has fields for neither type)
         del data['entry'][0]['messaging'][0]['sender']
         response = self.client.post(callback_url, json.dumps(data).replace('PAGE_ID', '1234'), content_type='application/json')
-        self.assertEqual(200, response.status_code)
-        self.assertEqual('{"status": ["Ignored opt-in, no user_ref or sender"]}', response.content)
+        self.assertContains(response, '{"status": ["Ignored opt-in, no user_ref or sender"]}')
 
         # ok, use a user_ref optin instead
         entry = json.loads(optin)
@@ -8081,8 +8073,7 @@ class FacebookTest(TembaTest):
         data = json.loads(FacebookTest.TEST_INCOMING)
         data['entry'][0]['messaging'][0] = json.loads(referral)
         response = self.client.post(callback_url, json.dumps(data), content_type='application/json')
-        self.assertEqual(200, response.status_code)
-        self.assertEqual('Msg Ignored for recipient id: PAGE_ID', response.content)
+        self.assertContains(response, 'Msg Ignored for recipient id: PAGE_ID')
 
         response = self.client.post(callback_url, json.dumps(data).replace('PAGE_ID', '1234'), content_type='application/json')
         self.assertEqual(200, response.status_code)
@@ -8105,8 +8096,7 @@ class FacebookTest(TembaTest):
         data['entry'][0]['messaging'][0].update(json.loads(referral))
         response = self.client.post(callback_url, json.dumps(data).replace('PAGE_ID', '1234'),
                                     content_type='application/json')
-        self.assertEqual(200, response.status_code)
-        self.assertEqual('{"status": ["Triggered flow for ref: not_handled"]}', response.content)
+        self.assertContains(response, '{"status": ["Triggered flow for ref: not_handled"]}')
 
         # check that the user started the flow
         contact1 = Contact.objects.get(org=self.org, urns__path='1122')
@@ -8128,8 +8118,7 @@ class FacebookTest(TembaTest):
         del data['entry'][0]['messaging'][0]['referral']
         data['entry'][0]['messaging'][0].update(json.loads(postback))
         response = self.client.post(callback_url, json.dumps(data).replace('PAGE_ID', '1234'), content_type='application/json')
-        self.assertEqual(200, response.status_code)
-        self.assertEqual('{"status": ["Referral posted with referral id: signup"]}', response.content)
+        self.assertContains(response, '{"status": ["Referral posted with referral id: signup"]}')
 
         # check that the user started the flow
         contact1 = Contact.objects.get(org=self.org, urns__path='1122')
@@ -8541,7 +8530,7 @@ class JiochatTest(TembaTest):
             self.assertTrue(ChannelLog.objects.filter(is_error=False).count(), 1)
             self.assertEqual(mock.call_count, 1)
 
-            self.assertEqual(channel_client.get_access_token(), 'ABC1234')
+            self.assertEqual(channel_client.get_access_token(), force_bytes('ABC1234'))
             self.assertEqual(mock.call_args_list[0][1]['data'], {'client_secret': u'app-secret',
                                                                  'grant_type': 'client_credentials',
                                                                  'client_id': u'app-id'})
@@ -8572,8 +8561,7 @@ class JiochatTest(TembaTest):
         response = self.client.get(callback_url +
                                    "?signature=%s&timestamp=%s&nonce=%s&echostr=SUCCESS" % (signature, timestamp,
                                                                                             nonce))
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.content, 'SUCCESS')
+        self.assertContains(response, 'SUCCESS')
         self.assertTrue(mock_refresh_access_token.called)
 
         mock_refresh_access_token.reset_mock()
@@ -8727,7 +8715,7 @@ class JiochatTest(TembaTest):
         self.assertEqual(ChannelLog.objects.filter(is_error=True).count(), 1)
 
         msg = Msg.objects.get()
-        self.assertEqual(response.content, "Msgs Accepted: %d" % msg.id)
+        self.assertContains(response, "Msgs Accepted: %d" % msg.id)
 
         # load our message
         self.assertIsNone(msg.contact.name)
@@ -8744,13 +8732,12 @@ class JiochatTest(TembaTest):
 
         mock_get.return_value = MockResponse(200, '{"nickname":"Shinonda"}')
         response = self.client.post(callback_url, json.dumps(data), content_type="application/json")
-        self.assertEqual(response.status_code, 200)
 
         self.assertEqual(ChannelLog.objects.all().count(), 1)
         self.assertEqual(ChannelLog.objects.filter(is_error=False).count(), 1)
 
         msg = Msg.objects.get()
-        self.assertEqual(response.content, "Msgs Accepted: %d" % msg.id)
+        self.assertContains(response, "Msgs Accepted: %d" % msg.id)
 
         # load our message
         self.assertEqual(msg.contact.name, "Shinonda")
@@ -8778,7 +8765,7 @@ class JiochatTest(TembaTest):
         self.assertEqual(ChannelLog.objects.filter(is_error=False).count(), 1)
 
         msg = Msg.objects.all().last()
-        self.assertEqual(response.content, "Msgs Accepted: %d" % msg.id)
+        self.assertContains(response, "Msgs Accepted: %d" % msg.id)
 
         # load our message
         self.assertEqual(msg.contact.name, "Shinonda")  # the name should not change to Kendrick
@@ -8795,12 +8782,11 @@ class JiochatTest(TembaTest):
 
         with AnonymousOrg(self.org):
             response = self.client.post(callback_url, json.dumps(data), content_type="application/json")
-            self.assertEqual(response.status_code, 200)
 
             self.assertEqual(ChannelLog.objects.all().count(), 0)
 
             msg = Msg.objects.get()
-            self.assertEqual(response.content, "Msgs Accepted: %d" % msg.id)
+            self.assertContains(response, "Msgs Accepted: %d" % msg.id)
 
             # load our message
             self.assertIsNone(msg.contact.name)
@@ -8843,7 +8829,7 @@ class JiochatTest(TembaTest):
                 self.assertEqual(ChannelLog.objects.all().first().response, '{"nickname": "Shinonda. \\u263a"}')
 
                 msg = Msg.objects.get()
-                self.assertEqual(response.content, "Msgs Accepted: %d" % msg.id)
+                self.assertContains(response, "Msgs Accepted: %d" % msg.id)
 
                 # load our message
                 self.assertEqual(msg.contact.get_urn(JIOCHAT_SCHEME).path, "1234")
@@ -8940,15 +8926,14 @@ class GlobeTest(TembaTest):
 
         msg = Msg.objects.get()
         self.assertEqual(msg.channel, self.channel)
-        self.assertEqual(response.content, "Msgs Accepted: %d" % msg.id)
+        self.assertContains(response, "Msgs Accepted: %d" % msg.id)
         Msg.objects.all().delete()
 
         # another valid post on the right address
         response = self.client.post(callback_url, json.dumps(data), content_type="application/json")
-        self.assertEqual(response.status_code, 200)
 
         msg = Msg.objects.get()
-        self.assertEqual(response.content, "Msgs Accepted: %d" % msg.id)
+        self.assertContains(response, "Msgs Accepted: %d" % msg.id)
 
         # load our message
         self.assertEqual(msg.contact.get_urn(TEL_SCHEME).path, "+639171234567")
@@ -9325,7 +9310,7 @@ class ViberPublicTest(TembaTest):
         data['message']['type'] = msg_type
         data['message'][payload_name] = payload_value
 
-        self.assertSignedRequest(json.dumps(data))
+        self.assertSignedRequest(force_bytes(json.dumps(data)))
 
         msg = Msg.objects.get()
         self.assertEqual(msg.text, assert_text)
@@ -9349,7 +9334,7 @@ class ViberPublicTest(TembaTest):
                 }
             }
 
-            response = self.assertSignedRequest(json.dumps(data), 400)
+            response = self.assertSignedRequest(force_bytes(json.dumps(data)), 400)
             self.assertContains(response, "Missing text or media in message in request body.", status_code=400)
             Msg.objects.all().delete()
 
@@ -9374,7 +9359,7 @@ class ViberPublicTest(TembaTest):
             "timestamp": 4987034606158369000,
             "message_token": 1481059480858
         }
-        self.assertSignedRequest(json.dumps(data))
+        self.assertSignedRequest(force_bytes(json.dumps(data)))
 
     def test_subscribed(self):
         data = {
@@ -9390,7 +9375,7 @@ class ViberPublicTest(TembaTest):
             },
             "message_token": 4912661846655238145
         }
-        self.assertSignedRequest(json.dumps(data))
+        self.assertSignedRequest(force_bytes(json.dumps(data)))
 
         # check that the contact was created
         contact = Contact.objects.get(org=self.org, urns__path='01234567890A=', urns__scheme=VIBER_SCHEME)
@@ -9402,13 +9387,13 @@ class ViberPublicTest(TembaTest):
             "user_id": "01234567890A=",
             "message_token": 4912661846655238145
         }
-        self.assertSignedRequest(json.dumps(data))
+        self.assertSignedRequest(force_bytes(json.dumps(data)))
         contact.refresh_from_db()
         self.assertTrue(contact.is_stopped)
 
         # use a user id we haven't seen before
         data['user_id'] = "01234567890B="
-        self.assertSignedRequest(json.dumps(data))
+        self.assertSignedRequest(force_bytes(json.dumps(data)))
 
         # should not create contacts we don't already know about
         self.assertIsNone(Contact.from_urn(self.org, URN.from_viber("01234567890B=")))
@@ -9428,7 +9413,7 @@ class ViberPublicTest(TembaTest):
                 },
                 "message_token": 4912661846655238145
             }
-            self.assertSignedRequest(json.dumps(data))
+            self.assertSignedRequest(force_bytes(json.dumps(data)))
 
             # check that the contact was created
             contact = Contact.objects.get(org=self.org, urns__path='01234567890A=', urns__scheme=VIBER_SCHEME)
@@ -9455,7 +9440,7 @@ class ViberPublicTest(TembaTest):
                 "api_version": 1
             }
         }
-        self.assertSignedRequest(json.dumps(data))
+        self.assertSignedRequest(force_bytes(json.dumps(data)))
 
     @override_settings(SEND_MESSAGES=True)
     def test_send(self):
@@ -9541,34 +9526,39 @@ class ViberPublicTest(TembaTest):
 
             Channel.send_message(dict_to_struct('MsgStruct', msg.as_task_json()))
 
+            json_data = dict(
+                auth_token='auth_token',
+                receiver="FXLP/JstS7kDuoiUGihkgA==",
+                text="Hello, world!",
+                type='text',
+                tracking_data=msg.id,
+                keyboard=dict(
+                    Type="keyboard",
+                    DefaultHeight=True,
+                    Buttons=[
+                        {
+                            "Text": "Yes",
+                            "ActionBody": "Yes",
+                            "ActionType": "reply",
+                            "TextSize": "regular"
+                        },
+                        {
+                            "Text": "No",
+                            "ActionBody": "No",
+                            "ActionType": "reply",
+                            "TextSize": "regular"
+                        }
+                    ]
+                )
+            )
+
             mock.assert_called_with('https://chatapi.viber.com/pa/send_message',
                                     headers={'Accept': u'application/json', u'User-agent': u'RapidPro'},
-                                    json=dict(
-                                        auth_token='auth_token',
-                                        receiver="FXLP/JstS7kDuoiUGihkgA==",
-                                        text="Hello, world!",
-                                        type='text',
-                                        tracking_data=msg.id,
-                                        keyboard=dict(
-                                            Type="keyboard",
-                                            DefaultHeight=True,
-                                            Buttons=[
-                                                {
-                                                    "Text": "Yes",
-                                                    "ActionBody": "Yes",
-                                                    "ActionType": "reply",
-                                                    "TextSize": "regular"
-                                                },
-                                                {
-                                                    "Text": "No",
-                                                    "ActionBody": "No",
-                                                    "ActionType": "reply",
-                                                    "TextSize": "regular"
-                                                }
-                                            ]
-                                        )
-                                    ),
+                                    json=ANY,
                                     timeout=5)
+
+            args, kwargs = mock.call_args
+            six.assertCountEqual(self, json_data, kwargs.get('json'))
 
             msg.refresh_from_db()
             self.assertEqual(msg.status, WIRED)
@@ -9683,13 +9673,16 @@ class FcmTest(TembaTest):
                 })
 
                 mock.assert_called_once_with('https://fcm.googleapis.com/fcm/send',
-                                             data=data,
+                                             data=ANY,
                                              headers={
                                                  'Content-Type': 'application/json',
                                                  'Authorization': 'key=123456789',
                                                  'User-agent': 'RapidPro'
                                              },
                                              timeout=5)
+
+                args, kwargs = mock.call_args
+                six.assertCountEqual(self, data, kwargs.get('data'))
 
                 self.clear_cache()
 
@@ -9739,13 +9732,16 @@ class FcmTest(TembaTest):
                 })
 
                 mock.assert_called_once_with('https://fcm.googleapis.com/fcm/send',
-                                             data=data,
+                                             data=ANY,
                                              headers={
                                                  'Content-Type': 'application/json',
                                                  'Authorization': 'key=123456789',
                                                  'User-agent': 'RapidPro'
                                              },
                                              timeout=5)
+
+                args, kwargs = mock.call_args
+                six.assertCountEqual(self, data, kwargs.get('data'))
 
                 self.clear_cache()
 
@@ -9785,13 +9781,18 @@ class FcmTest(TembaTest):
             })
 
             mock.assert_called_once_with('https://fcm.googleapis.com/fcm/send',
-                                         data=data,
+                                         data=ANY,
                                          headers={
                                              'Content-Type': 'application/json',
                                              'Authorization': 'key=123456789',
                                              'User-agent': 'RapidPro'
                                          },
                                          timeout=5)
+
+            # mock.assert_called_once_with only compares equality of arguments so it's not practical for use with
+            # complex arguments, which we need to check manually
+            args, kwargs = mock.call_args
+            six.assertCountEqual(self, data, kwargs.get('data'))
 
             self.clear_cache()
 
@@ -9803,7 +9804,7 @@ class CourierTest(TembaTest):
             self.channel.save()
 
             bob = self.create_contact("Bob", urn='tel:+12065551111')
-            incoming = self.create_msg(contact=bob, text="Hello", direction="I")
+            incoming = self.create_msg(contact=bob, text="Hello", direction="I", external_id="external-id")
 
             # create some outgoing messages for our channel
             msg1 = Msg.create_outgoing(self.org, self.admin, 'tel:+12065551111', "Outgoing 1",
@@ -9851,6 +9852,7 @@ class CourierTest(TembaTest):
             self.assertEqual(low_priority_msgs[0][0]['tps_cost'], 2)
             self.assertIsNone(low_priority_msgs[1][0]['attachments'])
             self.assertEqual(low_priority_msgs[1][0]['tps_cost'], 1)
+            self.assertEqual(low_priority_msgs[1][0]['response_to_external_id'], "external-id")
             self.assertIsNone(low_priority_msgs[2][0]['attachments'])