+++ temba/ivr/tests.py
@@ -329,7 +329,7 @@ class IVRTests(FlowFileTest):
     def test_ivr_recording_with_nexmo(self, mock_create_call, mock_create_application, mock_jwt):
         mock_create_application.return_value = dict(id='app-id', keys=dict(private_key='private-key'))
         mock_create_call.return_value = MockResponse(200, json.dumps(dict(uuid='12345')))
-        mock_jwt.return_value = 'Encoded data'
+        mock_jwt.return_value = b'Encoded data'
 
         # connect Nexmo
         self.org.connect_nexmo('123', '456', self.admin)
@@ -773,7 +773,7 @@ class IVRTests(FlowFileTest):
         # make sure we set submitOnHash to true nexmo
         self.assertContains(response, '"submitOnHash": true,')
 
-        self.assertContains(response, '"eventUrl": ["https://%s%s"]}]' % (self.channel.callback_domain, callback_url))
+        self.assertContains(response, '"eventUrl": ["https://%s%s"]' % (self.channel.callback_domain, callback_url))
 
     @patch('jwt.encode')
     @patch('requests.put')
@@ -782,7 +782,7 @@ class IVRTests(FlowFileTest):
     def test_expiration_hangup(self, mock_create_call, mock_create_application, mock_put, mock_jwt):
         mock_create_application.return_value = dict(id='app-id', keys=dict(private_key='private-key'))
         mock_create_call.return_value = MockResponse(200, json.dumps(dict(call=dict(uuid='12345'))))
-        mock_jwt.return_value = "Encoded data"
+        mock_jwt.return_value = b"Encoded data"
 
         request = MagicMock()
         request.body = json.dumps(dict(call_id='12345'))
@@ -1250,8 +1250,7 @@ class IVRTests(FlowFileTest):
 
         # no channel found for call handle
         response = self.client.post(reverse('ivr.ivrcall_handle', args=[call.pk]), dict())
-        self.assertEqual(400, response.status_code)
-        self.assertEqual('No channel found', response.content)
+        self.assertContains(response, 'No channel found', status_code=400)
 
     @patch('temba.ivr.clients.TwilioClient', MockTwilioClient)
     @patch('twilio.util.RequestValidator', MockRequestValidator)
@@ -1452,8 +1451,7 @@ class IVRTests(FlowFileTest):
         response = self.client.post(reverse('handlers.nexmo_call_handler', args=['answer', nexmo_uuid]),
                                     json.dumps(post_data), content_type="application/json")
 
-        self.assertEqual(404, response.status_code)
-        self.assertEqual('Channel not found for number: 250785551212', response.content)
+        self.assertContains(response, 'Channel not found for number: 250785551212', status_code=404)
 
         # no call object created
         self.assertFalse(IVRCall.objects.all())